+++ service/ecr/api.go
@@ -12662,7 +12662,9 @@ func (s *ScanNotFoundException) RequestID() string {
 	return s.RespMetadata.RequestID
 }
 
-// The details of a scanning repository filter.
+// The details of a scanning repository filter. For more information on how
+// to use filters, see Using filters (https://docs.aws.amazon.com/AmazonECR/latest/userguide/image-scanning.html#image-scanning-filters)
+// in the Amazon Elastic Container Registry User Guide.
 type ScanningRepositoryFilter struct {
 	_ struct{} `type:"structure"`